pyproject.toml
[tool.poetry]
name = "mcf-intelligence-platform"
version = "1.0.0"
description = "Hiring-market intelligence, explainable NLP search, and profile matching for MyCareersFuture.sg"
authors = ["Your Name <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
pydantic = "^2.5.0"
tenacity = "^8.2.0"
typer = {extras = ["all"], version = "^0.15.0"}
rich = "^13.7.0"
pandas = "^2.1.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
numpy = "^1.24.0"
cachetools = "^5.3.0"

[tool.poetry.group.ml]
optional = true

[tool.poetry.group.ml.dependencies]
torch = ">=2.0.0,<2.3.0"
sentence-transformers = "^2.2.0"
faiss-cpu = "^1.7.4"
scikit-learn = "^1.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-httpx = "^0.30.0"
ruff = "^0.4.0"

[tool.poetry.scripts]
mcf = "src.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = ["src/legacy/", "src/linkedin.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
